requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.4.0
certifi==2023.7.22
cffi==1.16.0
channels==4.1.0
chargily-epay-Django==0.0.2
chargily-epay-gateway==0.2
chargily-epay-gateway-django-plugin==0.1
chargily-epay-python==0.0.2
charset-normalizer==2.0.12
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.8
daphne==4.1.2
distlib==0.3.7
Django==4.2.4
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-ckeditor==6.7.0
django-cors-headers==1.1.0
django-credit-cards==0.4.1
django-cropper-image==1.0.5
django-filter==23.2
django-js-asset==2.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
filelock==3.12.4
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
kombu==5.3.7
Pillow==10.0.0
platformdirs==3.10.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.27.0
service-identity==24.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
vine==5.1.0
virtualenv==20.24.5
wcwidth==0.2.13
whitenoise==6.6.0
zope.interface==6.4.post2
